Refactoring: improved code quality.

+++ src/org/parabuild/ci/merge/finder/perforce/P4StartingChangeListFinder.java
@@ -76,7 +76,7 @@ public final class P4StartingChangeListFinder {
     ThreadUtils.checkIfInterrupted();
 
     if (log.isDebugEnabled()) log.debug("=========== end finding starting change list number ===========");
-    return Math.max(firstFromChangeList.intValue(), firstToChangeList.intValue());
+    return Math.max(firstFromChangeList, firstToChangeList);
   }
